requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
python_decouple
qrcode
requests
